% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamstr-bacon.R
\name{.StackIterations}
\alias{.StackIterations}
\title{Stack iterations}
\usage{
.StackIterations(x)
}
\arguments{
\item{x}{extracted bacon age models}
}
\description{
Stack iterations
}
\keyword{internal}
