% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamstr-bacon.R
\name{get_bacon_acc_rates}
\alias{get_bacon_acc_rates}
\title{Extract Bacon Accumulation Rates}
\usage{
get_bacon_acc_rates(hamstr_bacon_fit)
}
\value{

}
\description{
Extract Bacon Accumulation Rates
}
\keyword{internal}
