% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamstr-bacon.R
\name{plot.hamstr_bacon_fit}
\alias{plot.hamstr_bacon_fit}
\title{Title}
\usage{
\method{plot}{hamstr_bacon_fit}(object, type = c("default"), summarise = TRUE, ...)
}
\arguments{
\item{object}{a hamstr_bacon_fit object}
}
\value{
A ggplot object
}
\description{
Title
}
