% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamstr-bacon.R
\name{plot_bacon_age_models}
\alias{plot_bacon_age_models}
\title{Plot individual Bacon age models}
\usage{
plot_bacon_age_models(hamstr_bacon_fit, n.iter = 1000)
}
\arguments{
\item{hamstr_bacon_fit}{a hamstr_bacon_fit object}

\item{n.iter}{The number of iterations of the model to plot, defaults to
1000.}
}
\value{

}
\description{
Plot individual Bacon age models
}
\keyword{internal}
