% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamstr-bacon.R
\name{plot_summary_bacon_age_models}
\alias{plot_summary_bacon_age_models}
\title{Plot summarised Bacon age models}
\usage{
plot_summary_bacon_age_models(hamstr_bacon_fit)
}
\arguments{
\item{hamstr_bacon_fit}{}
}
\description{
Plot summarised Bacon age models
}
\keyword{internal}
