% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamstr-bacon.R
\name{predict.hamstr_bacon_fit}
\alias{predict.hamstr_bacon_fit}
\title{Interpolate Age Models at Given Depths}
\usage{
\method{predict}{hamstr_bacon_fit}(
  object,
  type = c("age_models", "acc_rates"),
  depth = c("modelled", "data"),
  ...
)
}
\arguments{
\item{object}{hamstr_fit object}

\item{type}{age models "age_models" or accumulation rates "acc_rates"}

\item{depth}{Defaults to "modelled", which returns the modelled depths.
"data" returns age models at the depths of the observations, or a numerical
 vector to specify depths. Accumulation rates are only returned at the 
 modelled depths.}
}
\value{

}
\description{
Method for generic function predict. Returns the posterior age
models interpolated to new depths given in depth.
}
