% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamstr-bacon.R
\name{return_bacon_age_mods}
\alias{return_bacon_age_mods}
\title{Extract and reconstruct bacon age models}
\usage{
return_bacon_age_mods(hamstr_bacon_fit)
}
\arguments{
\item{hamstr_bacon_fit}{a hamstr_bacon_fit object}
}
\description{
Extract and reconstruct bacon age models
}
\keyword{internal}
