% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamstr-bacon.R
\name{summarise_bacon_acc_rates}
\alias{summarise_bacon_acc_rates}
\title{Summarise Bacon Accumulation Rates}
\usage{
summarise_bacon_acc_rates(hamstr_bacon_fit)
}
\value{

}
\description{
Summarise Bacon Accumulation Rates
}
\keyword{internal}
