% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamstr-bacon.R
\name{summarise_bacon_age_models}
\alias{summarise_bacon_age_models}
\title{Summarise Bacon age models}
\usage{
summarise_bacon_age_models(hamstr_bacon_fit)
}
\arguments{
\item{hamstr_bacon_fit}{}
}
\description{
Summarise Bacon age models
}
\keyword{internal}
