% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamstr-bacon.R
\name{summarise_interpolated_bacon_age_models}
\alias{summarise_interpolated_bacon_age_models}
\title{Summarise interpolated Bacon age models}
\usage{
summarise_interpolated_bacon_age_models(interpolated_age_mods)
}
\arguments{
\item{interpolated_age_mods}{}
}
\description{
Summarise interpolated Bacon age models
}
\keyword{internal}
