% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamstr-bacon.R
\name{summary.hamstr_bacon_interpolated_ages}
\alias{summary.hamstr_bacon_interpolated_ages}
\title{Summarise hamstr_bacon Age Models}
\usage{
\method{summary}{hamstr_bacon_interpolated_ages}(object)
}
\description{
Summarise hamstr_bacon Age Models
}
