% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamstr-bacon.R
\name{get_bacon_acc_rates}
\alias{get_bacon_acc_rates}
\title{Extract Bacon Accumulation Rates}
\usage{
get_bacon_acc_rates(hamstr_bacon_fit, tau = 0, kern = c("U", "G", "BH"))
}
\arguments{
\item{hamstr_bacon_fit}{a hamstr_bacon_fit object}
}
\description{
Extract Bacon Accumulation Rates
}
\keyword{internal}
