% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamstr-bacon.R
\name{hamstr_bacon}
\alias{hamstr_bacon}
\title{Run Wrapped Bacon}
\usage{
hamstr_bacon(
  depth,
  obs_age,
  obs_err,
  thick = NULL,
  d.min = NA,
  d.max = NA,
  d.by = NA,
  plot.pdf = FALSE
)
}
\arguments{
\item{depth}{}

\item{obs_age}{}

\item{obs_err}{}

\item{thick}{Bacon will divide the core into sections of equal thickness specified by thick (default \code{thick=5}).}

\item{d.min}{Minimum depth of age-depth model (use this to extrapolate to depths higher than the top dated depth).}

\item{d.max}{Maximum depth of age-depth model (use this to extrapolate to depths below the bottom dated depth).}

\item{d.by}{Depth intervals at which ages are calculated. Defaults to \code{d.by=1}.}

\item{plot.pdf}{Produce a pdf file of the age-depth plot. Defaults to \code{plot.pdf=TRUE} after a Bacon run.}

\item{...}{}
}
\value{

}
\description{
Run Wrapped Bacon
}
