% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamstr-bacon.R
\name{hamstr_bacon}
\alias{hamstr_bacon}
\title{A Wrapper for rbacon::Bacon}
\usage{
hamstr_bacon(
  id = "default",
  depth,
  obs_age,
  obs_err,
  thick = 5,
  d.min = NA,
  d.max = NA,
  d.by = NULL,
  seed = NA,
  acc.shape = 1.5,
  acc.mean = 20,
  mem.strength = 10,
  mem.mean = 0.5,
  boundary = NA,
  hiatus.depths = NA,
  hiatus.max = 10000,
  add = c(),
  cc = 1,
  cc1 = "IntCal20",
  cc2 = "Marine20",
  cc3 = "SHCal20",
  cc4 = "ConstCal",
  ccdir = "",
  postbomb = 0,
  delta.R = 0,
  delta.STD = 0,
  t.a = 3,
  t.b = 4,
  normal = FALSE,
  suggest = TRUE,
  accept.suggestions = TRUE,
  reswarn = c(10, 200),
  ask = FALSE,
  slump = c(),
  remove = FALSE,
  ssize = 2000,
  th0 = c(),
  burnin = min(500, ssize),
  MinAge = c(),
  MaxAge = c(),
  plot.pdf = FALSE,
  close.connections = FALSE,
  verbose = FALSE,
  suppress.plots = TRUE
)
}
\arguments{
\item{depth}{A vector of depths}

\item{obs_age}{A vector of observed ages}

\item{obs_err}{A vector of errors on the observed ages}
}
\value{

}
\description{
Wraps the Bacon function from rbacon so that it can be used in a
more typical "R" way. Returns age-depth models in a format to match hamstr
output. Most Bacon functionality is accessible, for example hiatuses and
boundaries, different calibration curves, postbomb curves.
}
\examples{
\dontrun{
hb1 <- hamstr_bacon(id = "sdf", 
                    depth = MSB2K$depth,
                    obs_age = MSB2K$age,
                    obs_err = MSB2K$error 
                    )

plot(hb1)
predict(hb1)
summary(hb1)
}
}
