% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamstr-bacon.R
\name{interpolate_bacon_age_models}
\alias{interpolate_bacon_age_models}
\title{Title}
\usage{
interpolate_bacon_age_models(hamstr_bacon_fit, new_depth)
}
\arguments{
\item{hamstr_bacon_fit}{}

\item{new_depth}{}
}
\value{

}
\description{
Title
}
