% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamstr-bacon.R
\name{plot.hamstr_bacon_fit}
\alias{plot.hamstr_bacon_fit}
\title{Title}
\usage{
\method{plot}{hamstr_bacon_fit}(
  hamstr_bacon_fit,
  type = c("default", "age_models", "acc_rates"),
  summarise = TRUE,
  ...
)
}
\arguments{
\item{hamstr_bacon_fit}{a hamstr_bacon_fit object}
}
\value{
A ggplot object
}
\description{
Title
}
