% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamstr-bacon.R
\name{plot_bacon_acc_rates}
\alias{plot_bacon_acc_rates}
\title{Plot summarised Bacon accumulation rates}
\usage{
plot_bacon_acc_rates(
  hamstr_bacon_fit,
  units = c("depth_per_time", "time_per_depth")
)
}
\arguments{
\item{hamstr_bacon_fit}{a hamstr_bacon_fit object}
}
\description{
Plot summarised Bacon accumulation rates
}
\keyword{internal}
