% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamstr-bacon.R
\name{plot_hamstr_bacon_fit}
\alias{plot_hamstr_bacon_fit}
\title{Title}
\usage{
plot_hamstr_bacon_fit(hamstr_bacon_fit, summarise = TRUE, n.iter = 1000, ...)
}
\arguments{
\item{hamstr_bacon_fit}{a hamstr_bacon_fit object}
}
\description{
Title
}
\keyword{internal}
