% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamstr-bacon.R
\name{predict.hamstr_bacon_fit}
\alias{predict.hamstr_bacon_fit}
\title{Interpolate Age Models at Given Depths}
\usage{
\method{predict}{hamstr_bacon_fit}(
  object,
  type = c("age_models", "acc_rates"),
  depth = c("modelled", "data"),
  ...
)
}
\description{
Method for generic function predict. Returns the posterior age
models interpolated to new depths given in depth.
}
