% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamstr-bacon.R
\name{summary.hamstr_bacon_fit}
\alias{summary.hamstr_bacon_fit}
\title{Summarise hamstr_bacon Age Models}
\usage{
\method{summary}{hamstr_bacon_fit}(object, type = c("default", "age_models", "acc_rates"), ...)
}
\description{
Summarise hamstr_bacon Age Models
}
